+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org)
 ### Added
 - Added aliases for `data_service`: `decision_service_ping` and `decision_service_post` ([ch23031](https://app.clubhouse.io/active-prospect/story/23031/integration-wrappers-for-tfds-in-ac-vc-pricing))
 
+## Fixed
+- Added validation of TrustedForm Data Service token environment variable
+
 ## [1.14.2] - 2021-04-14
 ### Fixed
 - Added additional error handling to Claim integration